requirements.txt
Flask==3.0.0
google-genai>=1.0.0
python-dotenv==1.0.1
reportlab==4.0.9
